feat: Add EmployeeReportController and report services that depend on employeeId
In order for requests to be possible for particular employees, the
EmployeeReportController was creates, along with corresponding services.
The EmployeeReportController's uris start with "employees/{employeeId}/reports"
and will expose services similar to the ReportController, but with the
additional requirement, that all report operations are performed on reports
belonging to a particular employee (identified with the parameter "employeeId").